Код адаптирован под росовские паеты и убраны алгос ошибки

+++ Core/Inc/Thruster.h
@@ -4,4 +4,4 @@
 #define kpercsec 0.5
 void Thruster_Init(void);
 void Thruster_Stop(void);
-void Thruster_Set_Speed(uint16_t *speed);
+void Thruster_Set_Speed(uint8_t *speed);
